backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
openpyxl==3.1.5
sqlalchemy==2.0.41
PyMySQL==1.1.1
python-dotenv==1.0.1

